[gw] Add feature sheet scrolling and Collections list (WIP)
At this point in the implementation, you can scroll in both medium and
large detents. It should work with both mouse and touch scrolling and
dragging interactions.

There's a (temporary) bug with the medium detent in that the bottom part
of the tab detail will be 'cut off', you can't scroll to it. That's
because the sheet _and_ the tab detail area both have a fixed size (the
former is `95vh`)... so some of the content is hidden off-screen.

I'm still deciding if 'scroll down' gestures in medium detent should
cause the card to expand to large detent, OR if it does in fact make
sense in this app to be able to scroll and interact with the tab detail
area in a medium detent -- i.e., while being able to see the map on the
top half of your screen, rather than only being able to interact if the
map is 95% occluded by the sheet.

Because the header area is large enough (at least 2x the minimum 44px
touch target size), it seems OK to only allow sheet resizing by dragging
the header.

If we decide to keep allowing scrolling and interaction in medium
detent, I might see if it's possible to dynamically resize
the height of either the sheet or the tab detail content, performantly
with a CSS transform. If that isn't a technically feasible solution, we
might have no choice but to go the other route.

Sheets are fun. :D